fixed create database
