+++ sdk/cosmos/src/clients/cosmos_struct.rs
@@ -75,11 +75,11 @@ impl CosmosUriBuilder for DefaultCosmosUri {
 
 #[derive(Debug, Clone, Default)]
 pub struct ChinaCosmosUri {
-    uri: String,
+    pub(crate) uri: String,
 }
 
 impl ChinaCosmosUri {
-    fn new(account: &str) -> ChinaCosmosUri {
+    pub(crate) fn new(account: &str) -> ChinaCosmosUri {
         ChinaCosmosUri {
             uri: format!("https://{}.documents.azure.cn", account),
         }
@@ -94,7 +94,7 @@ impl CosmosUriBuilder for ChinaCosmosUri {
 
 #[derive(Debug, Clone, Default)]
 pub struct CustomCosmosUri {
-    uri: String,
+    pub(crate) uri: String,
 }
 
 impl CosmosUriBuilder for CustomCosmosUri {
@@ -123,62 +123,6 @@ impl<'a, CUB> ClientBuilder<'a, CUB, HTTPClientNotAssigned>
 where
     CUB: CosmosUriBuilder,
 {
-    pub fn new_china<IntoCowStr>(
-        account: IntoCowStr,
-        auth_token: AuthorizationToken,
-    ) -> Result<ClientBuilder<'a, ChinaCosmosUri, HTTPClientNotAssigned>, AzureError>
-    where
-        IntoCowStr: Into<Cow<'a, str>>,
-    {
-        let account = account.into();
-        let cosmos_uri_builder = ChinaCosmosUri::new(account.as_ref());
-
-        Ok(ClientBuilder {
-            http_client: None,
-            p_http_client_to_assign: PhantomData {},
-            account,
-            auth_token,
-            cosmos_uri_builder,
-        })
-    }
-
-    pub fn new_custom<IntoCowStr>(
-        account: IntoCowStr,
-        auth_token: AuthorizationToken,
-        uri: String,
-    ) -> Result<ClientBuilder<'a, CustomCosmosUri, HTTPClientNotAssigned>, AzureError>
-    where
-        IntoCowStr: Into<Cow<'a, str>>,
-    {
-        Ok(ClientBuilder {
-            http_client: None,
-            p_http_client_to_assign: PhantomData {},
-            account: account.into(),
-            auth_token,
-            cosmos_uri_builder: CustomCosmosUri { uri },
-        })
-    }
-
-    pub fn new_emulator(
-        address: &str,
-        port: u16,
-    ) -> Result<ClientBuilder<CustomCosmosUri, HTTPClientNotAssigned>, AzureError> {
-        //Account name: localhost:<port>
-        //Account key: C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==
-        let auth_token = AuthorizationToken::new_master(
-            "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==",
-        ).unwrap();
-        Ok(ClientBuilder {
-            http_client: None,
-            p_http_client_to_assign: PhantomData {},
-            account: Cow::Owned(format!("{}:{}", address, port)),
-            auth_token,
-            cosmos_uri_builder: CustomCosmosUri {
-                uri: format!("https://{}:{}", address, port),
-            },
-        })
-    }
-
     pub fn with_http_client(
         self,
         http_client: Arc<Box<dyn HttpClient>>,
@@ -221,9 +165,9 @@ impl<'a, CUB> CosmosClient for CosmosStruct<'a, CUB>
 where
     CUB: CosmosUriBuilder + Debug,
 {
-    //fn create_database(&self) -> requests::CreateDatabaseBuilder<'_, No> {
-    //    requests::CreateDatabaseBuilder::new(self)
-    //}
+    fn create_database(&self) -> requests::CreateDatabaseBuilder<'_, No> {
+        requests::CreateDatabaseBuilder::new(self)
+    }
 
     //fn list_databases(&self) -> requests::ListDatabasesBuilder<'_> {
     //    requests::ListDatabasesBuilder::new(self)
