+++ sdk/cosmos/src/requests/mod.rs
@@ -1,5 +1,5 @@
 //mod create_collection_builder;
-//mod create_database_builder;
+mod create_database_builder;
 //mod create_document_builder;
 //mod create_or_replace_trigger_builder;
 //mod create_or_replace_user_defined_function_builder;
@@ -43,7 +43,7 @@ mod get_database_builder;
 //mod replace_stored_procedure_builder;
 //mod replace_user_builder;
 //pub use self::create_collection_builder::CreateCollectionBuilder;
-//pub use self::create_database_builder::CreateDatabaseBuilder;
+pub use self::create_database_builder::CreateDatabaseBuilder;
 //pub use self::create_document_builder::CreateDocumentBuilder;
 //pub use self::create_or_replace_trigger_builder::CreateOrReplaceTriggerBuilder;
 //pub use self::create_or_replace_user_defined_function_builder::CreateOrReplaceUserDefinedFunctionBuilder;
